pyproject.toml
[project]
name = "erudit-backend"
version = "0.0.1"
description = "This project is the backend service for Erudit. It provides APIs and handles data processing for the Erudit application."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "advanced-alchemy>=1.6.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "passlib[argon2]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "ruff>=0.12.12",
    "structlog>=25.4.0",
    "uvicorn>=0.35.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

exclude = [
    ".venv",
    "migrations"
]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
